requirements.txt
einops>=0.6.1
huggingface-hub>=0.20.3
numpy>=1.22.3
pandas>=1.4.1
transformers>=4.17.0
wandb>=0.16.3
torch>=2.1.0
torchtyping>=0.1.4
tqdm>=4.65.0
prettytable>=3.9.0
python-dotenv>=1.0.0
git-lfs>=1.6
datasets>=2.17.0
